requirements.txt
apache-airflow-providers-postgres
pandas
faker
dbt-core==1.8.2
dbt-postgres==1.8.2